+++ luqum/elasticsearch/visitor.py
@@ -48,12 +48,13 @@ class ElasticsearchQueryBuilder(LuceneTreeVisitorV2):
     CONTEXT_FIELD_PREFIX = "field_prefix"
 
     def __init__(self, default_operator=SHOULD, default_field='text',
-                 not_analyzed_fields=None, nested_fields=None, object_fields=None,
+                 not_analyzed_fields=None, nested_fields=None, object_fields=None, sub_fields=None,
                  field_options=None, match_word_as_phrase=False):
         """
         :param default_operator: to replace blank operator (MUST or SHOULD)
         :param default_field: to search
         :param not_analyzed_fields: field that are not analyzed in ES
+          (do not forget to include eventual sub fields)
         :param nested_fields: dict contains fields that are nested in ES
             each nested fields contains
             either a dict of nested fields
@@ -74,6 +75,8 @@ class ElasticsearchQueryBuilder(LuceneTreeVisitorV2):
         :param object_fields: list containing full qualified names of object fields.
           You may also use a spec similar to the one used for nested_fields.
           None, will accept all non nested fields as object fields.
+        :param sub_fields: list containing full qualified names of sub fields.
+          None, will accept all non nested fields or object fields as sub fields.
         :param dict field_options: allows you to give defaults options for each fields.
           They will be applied unless, overwritten by generated parameters.
           For match query, the `type` parameter modifies the query type.
@@ -99,6 +102,7 @@ class ElasticsearchQueryBuilder(LuceneTreeVisitorV2):
             k.rsplit(".", 1)[0]
             for k in flatten_nested_fields_specs(self.nested_fields))
         self.object_fields = self._normalize_object_fields(object_fields)
+        self.sub_fields = sub_fields
         self.field_options = field_options or {}
         self.default_operator = default_operator
         self.default_field = default_field
@@ -108,7 +112,10 @@ class ElasticsearchQueryBuilder(LuceneTreeVisitorV2):
             field_options=self.field_options,
         )
         self.nesting_checker = CheckNestedFields(
-            nested_fields=self.nested_fields, object_fields=self.object_fields)
+            nested_fields=self.nested_fields,
+            object_fields=self.object_fields,
+            sub_fields=self.sub_fields,
+        )
         if match_word_as_phrase:
             warnings.warn(
                 "match_word_as_phrase is a transient option " +